feat: add data script
